[TG-TECH][refactor] changing folder structure for components
