--- src/app/components/device-list/device-list.scss
@@ -1,40 +1,0 @@
-@mixin sorting {
-    width: 0;
-    height: 0;
-    border-left: 5px solid transparent;
-    border-right: 5px solid transparent;
-}
-
-.sub-header {
-    margin: 25px auto;
-
-    #page-title {
-        font-size:36px;
-    }
-
-    .actions {
-      vertical-align: middle;
-      line-height: 54px;
-    }
-}
-.sortable {
-    cursor: pointer;
-}
-
-.active {
-    .up {
-        @include sorting;
-        border-bottom: 5px solid black;
-    }
-
-    .down {
-        @include sorting;
-        border-top: 5px solid black;
-    }
-}
-
-.empty-list {
-    &::before {
-        background-image: url(./assets/empty-sensors-list.png);
-    }
-}
