+++ src/app/dashboard/base-sensor/base-sensor.test.js
