+++ src/app/dashboard/base-sensor/index.js
@@ -1,0 +1,2 @@
+import BaseSensor from './base-sensor';
+export default BaseSensor;
