+++ src/app/dashboard/dashboard-editor/dashboard-editor.js
@@ -4,8 +4,8 @@ import {Router, RouterLink} from 'angular2/router';
 import template from './dashboard-editor.html';
 import style from '../style.scss';
 import DashboardService from '../dashboard.service';
-import { SensorWidget } from '../../components/sensor-widget';
-import { SensorExecutorWidget } from '../../components/sensor-executor-widget';
+import { SensorWidget } from '../sensor-widget';
+import { SensorExecutorWidget } from '../sensor-executor-widget';
 import { Dragula, DragulaService } from 'ng2-dragula/ng2-dragula';
 
 const selector = 'dashboard-editor';
