+++ src/app/dashboard/sensor-executor-widget/index.js
