+++ src/app/dashboard/sensor-executor-widget/sensor-executor-widget.html
