+++ src/app/dashboard/sensor-status-widget/index.js
