+++ src/app/dashboard/sensor-status-widget/sensor-status-widget.component.js
@@ -1,6 +1,6 @@
 import { Component } from 'angular2/core';
 
-import BaseSensor from '../shared/base-sensor';
+import BaseSensor from '../base-sensor';
 import template from './sensor-status-widget.html';
 import baseStyles from '../shared/sensor-widget.scss';
 import styles from './sensor-status-widget.scss';
