+++ src/app/dashboard/sensor-status-widget/sensor-status-widget.scss
