+++ src/app/dashboard/sensor-widget/index.js
