+++ src/app/dashboard/sensor-widget/sensor-widget.component.js
@@ -1,6 +1,6 @@
 import { Component } from 'angular2/core';
 
-import BaseSensor from '../shared/base-sensor';
+import BaseSensor from '../base-sensor';
 import template from './sensor-widget.html';
 import style from '../shared/sensor-widget.scss';
