+++ src/app/dashboard/sensor-widget/sensor-widget.html
