+++ src/app/dashboard/shared/sensor-widget.scss
