+++ src/app/dashboard/shared/sensor-widget.service.js
