+++ src/app/dashboard/shared/sensor-widget.service.test.js
