+++ src/app/devices/details/index.js
