+++ src/app/devices/details/sensor-detail.component.js
