+++ src/app/devices/details/sensor-detail.component.test.js
