+++ src/app/devices/details/sensor-detail.css
