+++ src/app/devices/details/sensor-detail.html
