+++ src/app/devices/details/sensor-detail.service.test.js
