+++ src/app/devices/details/sensor.js
