+++ src/app/devices/index.js
@@ -1,0 +1,4 @@
+import { DeviceList } from './list';
+import { SensorDetail } from './details';
+
+export { DeviceList, SensorDetail };
