+++ src/app/devices/list/assets/empty-sensors-list.png
