+++ src/app/devices/list/device-list.component.js
@@ -64,7 +64,7 @@ export class DeviceList {
     }
 
     removeSensor(item) {
-        if (!window.confirm(confirmQuestion)) {
+        if (!window.confirm(confirmQuestion)) { // eslint-disable-line no-alert
             return;
         }
