+++ src/app/devices/list/device-list.component.test.js
