+++ src/app/devices/list/device-list.html
