+++ src/app/devices/list/device-list.scss
