+++ src/app/devices/list/index.js
