--- src/app/sensors/details/sensor-detail.test.js
@@ -1,9 +1,0 @@
-import {SensorDetail} from './sensor-detail.component';
-
-describe('SensorEdit', () => {
-    describe('#constructor', () => {
-        it('should be defined', () => {
-            expect(SensorDetail).toBeDefined();
-        });
-    });
-});
