+++ src/app/shared/device-list/device-list.service.test.js
